Update bitshortener.py
Incorporating a requests.Session object into the BitShortener class is a great way to make HTTP requests more efficient. This is because a Session object uses persistent HTTP connections, which is beneficial when making multiple requests to the same host, as it reduces the overhead of establishing a new connection for each request.